dewPoint, absolute humidity and location

+++ app/src/main/java/fi/carterm/clearskiesweather/models/WeatherData.kt
@@ -11,8 +11,8 @@ data class WeatherData(
     val humidity: Float,
     val pressure: Float,
     val light: Float,
-    val abshumidity: Float,
-    val dewpoint: Float,
+    val abshumidity: Double,
+    val dewpoint: Double,
     val latitude: Double,
-    val longitude: Double,
+   val longitude: Double,
 )
